Update and rename Cpp/Autocomplete.cpp to cpp/Autocomplete.cpp
added comments and indentation to main()
